Have producer re-connect in go routine
Currently, re-connects end up blocking the processing of the `eventChan`
on the producer. This causes the `eventChan` to get full and block
enqueueing of messages. This makes it difficult in cases where we want
to make sure we don't block on `producer.SendAsync`.

I am not sure if this has negative effects, but it seems like we still
would be okay emitting events into the producer write buffer while
disconnected, especially as it isn't guranteed that the method succeeds