Add support for not blocking on SendAsync
In some cases, it may be useful to not have SendAsync ever block.

For example, if you have code around a producer that takes care of retries
natively. To support this, we add new option `NeverBlockOnSendAsync`.

In that mode, if the eventsChan or maxPendingMessages buffer fills up,
the callback immediately returns with an error that indicates those
states.